Move image manipulation stuff to own subpackage
Wow I am starting to see the light!!!11elf
